+++ src/main/java/net/geforcemods/securitycraft/api/CustomizableSCTE.java
@@ -6,8 +6,6 @@ import net.geforcemods.securitycraft.misc.EnumModuleType;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.NonNullList;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.TextComponentString;
 
 /**
  * Extend this class in your TileEntity to make it customizable. You will be able to modify it with the various modules in
@@ -34,11 +32,6 @@ public abstract class CustomizableSCTE extends TileEntityNamed implements IModul
 		writeModuleStates(tag);
 		writeOptions(tag);
 		return tag;
-	}
-
-	@Override
-	public ITextComponent getDefaultName() {
-		return new TextComponentString("Customize");
 	}
 
 	@Override